requirements.txt
opencv-python
numpy
PyAutoGUI
pynput
Pillow
dearpygui
pytesseract